Update submodules to recent master
Latest versions of the ilk-generator, ilk-compiler and
robot-model-tools that should work together.

Note that the robot-model-tools is now listed in the project
dependencies of the ilk-generator (in pyproject.toml), but
for now I decided to keep the source tree clone in the
submodule. Just in case we will have to point to a development
version of the tools, which is not in pypi.

NOTE: the tests in this repo are completely outdated and
need to be reworked.